test for mcinversion version2
